Create produce_mapping_file.py
produce_mapping_file.py takes as arguments a folder of sequencing files that have been renamed according to sample labels, and an output location and Mapping file name. It creates a Mapping.txt file containing for use with Lotus2 to be integrated with a metadata file. The mapping file pairs R1 and R2 files into a column (comma-separated), and assigns whether samples are true samples or controls.